requirements.txt
anyascii==0.3.1
appnope==0.1.3
asgiref==3.5.2
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.0.1
click==8.1.3
comm==0.1.2
confection==0.0.4
contractions==0.1.73
cymem==2.0.7
debugpy==1.6.6
decorator==5.1.1
Django==4.0
django-cors-headers==3.13.0
djangorestframework==3.13.0
djangorestframework-simplejwt==5.2.0
executing==1.2.0
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.21.1
ipython==8.9.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.0.2
jupyter_core==5.2.0
langcodes==3.3.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
msgpack==1.0.4
murmurhash==1.0.9
mysql-connector-python==8.0.32
mysqlclient==2.1.1
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.24.1
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
preshed==3.0.8
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyahocorasick==2.0.0
pydantic==1.10.4
Pygments==2.14.0
PyJWT==2.4.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.2.1
pyzmq==25.0.0
redis==4.4.2
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
smart-open==6.3.0
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.2
srsly==2.4.5
stack-data==0.6.2
textsearch==0.0.24
thinc==8.1.7
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.14
vaderSentiment==3.3.2
wasabi==1.1.1
wcwidth==0.2.6
wordninja==2.0.0
xgboost==1.7.3
zipp==3.12.0
